switched bushfiretute to use Dataserver time; removed Global time

+++ examples/bushfire-tutorial/src/main/java/io/github/agentsoz/bushfiretute/datacollection/ScenarioTwoData.java
@@ -64,7 +64,7 @@ public static void writeConnectToDepTimesToFile() {
 			writer.println("id" + "\t" + "deptime");
 	        for(int i = 0 ; i < connectToAgents.size() ; i++){
 	        	String id = (String)connectToAgents.get(i) ;
-	        	long time  = (long)connectToDepTimes.get(i);
+	        	double time  = (double)connectToDepTimes.get(i);
 	        	writer.println(id + "\t" + time);
 	            }
